+++ Riot/Modules/TabBar/MasterTabBarController.m
@@ -27,6 +27,9 @@
 #import "MXRoom+Riot.h"
 #import "MXSession+Riot.h"
 
+#import "SettingsViewController.h"
+#import "SecurityViewController.h"
+
 #import "Riot-Swift.h"
 
 @interface MasterTabBarController ()
@@ -66,6 +69,8 @@
 
 @property(nonatomic,getter=isHidden) BOOL hidden;
 
+@property(nonatomic) BOOL reviewSessionAlertHasBeenDisplayed;
+
 @end
 
 @implementation MasterTabBarController
@@ -74,6 +79,9 @@
 {
     [super viewDidLoad];
     // Do any additional setup after loading the view, typically from a nib.
+    
+    // Note: UITabBarViewController shoud not be embed in a UINavigationController (https://github.com/vector-im/riot-ios/issues/3086)
+    [self vc_removeBackTitle];
 
     // Retrieve the all view controllers
     _homeViewController = self.viewControllers[TABBAR_HOME_INDEX];
@@ -206,6 +214,8 @@
             
             [childViewControllers removeAllObjects];
         }
+        
+        [self presentReviewSessionsAlertIfNeeded];
     }
     
     if (unifiedSearchViewController)
@@ -339,6 +349,9 @@
             
             // Prepare data sources and return
             [self initializeDataSources];
+            
+            // Add matrix sessions observer on first added session
+            [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(onMatrixSessionStateDidChange:) name:kMXSessionStateDidChangeNotification object:nil];
             return;
         }
         else
@@ -387,6 +400,8 @@
 - (void)onMatrixSessionStateDidChange:(NSNotification *)notif
 {
     [self refreshTabBarBadges];
+    
+    [self presentReviewSessionsAlertIfNeeded];
 }
 
 - (void)showAuthenticationScreen
@@ -397,6 +412,8 @@
     if (!self.authViewController && !isAuthViewControllerPreparing)
     {
         isAuthViewControllerPreparing = YES;
+        
+        [self resetReviewSessionsFlags];
         
         [[AppDelegate theDelegate] restoreInitialDisplay:^{
             
@@ -974,6 +991,147 @@
     
     [currentAlert mxk_setAccessibilityIdentifier: @"HomeVCUseAnalyticsAlert"];
     [self presentViewController:currentAlert animated:YES completion:nil];
+}
+
+#pragma mark - Review session
+
+- (void)presentReviewSessionsAlertIfNeeded
+{
+    MXSession *mainSession = self.mxSessions.firstObject;
+    
+    if (!(self.viewLoaded
+          && mainSession.state >= MXSessionStateStoreDataReady
+          && mainSession.crypto.crossSigning))
+    {
+        return;
+    }
+    
+    switch (mainSession.crypto.crossSigning.state) {
+        case MXCrossSigningStateCrossSigningExists:
+            [self presentVerifyCurrentSessionAlertIfNeededWithSession:mainSession];
+            break;
+        case MXCrossSigningStateCanCrossSign:
+            [self presentReviewUnverifiedSessionsAlertIfNeededWithSession:mainSession];
+            break;
+        default:
+            break;
+    }
+}
+
+- (void)presentVerifyCurrentSessionAlertIfNeededWithSession:(MXSession*)session
+{
+    if (RiotSettings.shared.hideVerifyThisSessionAlert || self.reviewSessionAlertHasBeenDisplayed)
+    {
+        return;
+    }
+    
+    self.reviewSessionAlertHasBeenDisplayed = YES;
+    [self presentVerifyCurrentSessionAlertWithSession:session];
+}
+
+- (void)presentVerifyCurrentSessionAlertWithSession:(MXSession*)session
+{
+    [currentAlert dismissViewControllerAnimated:NO completion:nil];
+    
+    UIAlertController *alert = [UIAlertController alertControllerWithTitle:NSLocalizedStringFromTable(@"key_verification_self_verify_current_session_alert_title", @"Vector", nil)
+                                                                   message:NSLocalizedStringFromTable(@"key_verification_self_verify_current_session_alert_message", @"Vector", nil)
+                                                            preferredStyle:UIAlertControllerStyleAlert];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"key_verification_self_verify_current_session_alert_validate_action", @"Vector", nil)
+                                              style:UIAlertActionStyleDefault
+                                            handler:^(UIAlertAction * action) {
+                                                [[AppDelegate theDelegate] presentCompleteSecurityForSession:session];
+                                            }]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"later", @"Vector", nil)
+                                              style:UIAlertActionStyleCancel
+                                            handler:nil]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"do_not_ask_again", @"Vector", nil)
+                                              style:UIAlertActionStyleDestructive
+                                            handler:^(UIAlertAction * action) {
+                                                RiotSettings.shared.hideVerifyThisSessionAlert = YES;
+                                            }]];
+    
+    
+    [self presentViewController:alert animated:YES completion:nil];
+    
+    currentAlert = alert;
+}
+
+- (void)presentReviewUnverifiedSessionsAlertIfNeededWithSession:(MXSession*)session
+{
+    if (RiotSettings.shared.hideReviewSessionsAlert || self.reviewSessionAlertHasBeenDisplayed)
+    {
+        return;
+    }
+    
+    NSArray<MXDeviceInfo*> *devices = [session.crypto.store devicesForUser:session.myUserId].allValues;
+    
+    BOOL isUserHasOneUnverifiedDevice = NO;
+    
+    for (MXDeviceInfo *device in devices)
+    {
+        if (!device.trustLevel.isCrossSigningVerified)
+        {
+            isUserHasOneUnverifiedDevice = YES;
+            break;
+        }
+    }
+    
+    if (isUserHasOneUnverifiedDevice)
+    {
+        self.reviewSessionAlertHasBeenDisplayed = YES;
+        [self presentReviewUnverifiedSessionsAlertWithSession:session];
+    }
+}
+
+- (void)presentReviewUnverifiedSessionsAlertWithSession:(MXSession*)session
+{
+    [currentAlert dismissViewControllerAnimated:NO completion:nil];
+    
+    UIAlertController *alert = [UIAlertController alertControllerWithTitle:NSLocalizedStringFromTable(@"key_verification_self_verify_unverified_sessions_alert_title", @"Vector", nil)
+                                                                   message:NSLocalizedStringFromTable(@"key_verification_self_verify_unverified_sessions_alert_message", @"Vector", nil)
+                                                            preferredStyle:UIAlertControllerStyleAlert];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"key_verification_self_verify_unverified_sessions_alert_validate_action", @"Vector", nil)
+                                              style:UIAlertActionStyleDefault
+                                            handler:^(UIAlertAction * action) {
+                                                [self showSettingsSecurityScreenForSession:session];
+                                            }]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"later", @"Vector", nil)
+                                              style:UIAlertActionStyleCancel
+                                            handler:nil]];
+    
+    [alert addAction:[UIAlertAction actionWithTitle:NSLocalizedStringFromTable(@"do_not_ask_again", @"Vector", nil)
+                                              style:UIAlertActionStyleDestructive
+                                            handler:^(UIAlertAction * action) {
+                                                RiotSettings.shared.hideReviewSessionsAlert = YES;
+                                            }]];
+    
+    
+    [self presentViewController:alert animated:YES completion:nil];
+    
+    currentAlert = alert;
+}
+
+- (void)showSettingsSecurityScreenForSession:(MXSession*)session
+{
+    SettingsViewController *settingsViewController = [SettingsViewController instantiate];
+    [settingsViewController loadViewIfNeeded];
+    SecurityViewController *securityViewController = [SecurityViewController instantiateWithMatrixSession:session];
+    
+    [[AppDelegate theDelegate] restoreInitialDisplay:^{
+        self.navigationController.viewControllers = @[self, settingsViewController, securityViewController];
+    }];
+}
+
+- (void)resetReviewSessionsFlags
+{
+    self.reviewSessionAlertHasBeenDisplayed = NO;
+    RiotSettings.shared.hideVerifyThisSessionAlert = NO;
+    RiotSettings.shared.hideReviewSessionsAlert = NO;
 }
 
 #pragma mark - UITabBarDelegate